requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiounittest==1.4.2
anyio==3.6.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.5
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
python-telegram-bot==20.2
rfc3986==1.5.0
sniffio==1.3.0
typing-extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
